finished interface with manager
I made a lot of changes this time. I added a few more functions in the
octree so the aabb's move and can change nodes. I improved accellerate
slightly and changed collisionDetection back to returning a bool. I
fixed the octree so it doesn't just check the first collision when we do
the collision detection teh way we had it before.